+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add relation between the `.h`, `.cxx`, `.ixx` and `.hxx` files in `projectile-other-file-alist`.
 * Add support to specify project name either via `.dir-locals.el` or by providing a customized `projectile-project-name-function'.
 * Add a command to switch between open projects (`projectile-switch-open-project`).
+* Add a command to edit the .dir-locals.el file of the project (`projectile-edit-dir-locals`).
 
 ### Changes